refactor: #111 - Simplifica layouts
