+++ client/src/app/(main)/layout.tsx
@@ -4,11 +4,9 @@ export default function MainLayout({
   children: React.ReactNode;
 }>) {
   return (
-    <div>
-      <div className="flex h-[100vh] w-full">
-        <aside className="bg-action-100 px-[16px] py-[24px]">EA</aside>
-        <div className="grow overflow-y-scroll p-[24px]">{children}</div>
-      </div>
+    <div className="flex h-[100vh]">
+      <aside className="bg-action-100 px-[16px] py-[24px]">EA</aside>
+      <div className="grow overflow-y-scroll p-[24px]">{children}</div>
     </div>
   );
 }
