--- client/src/app/(main)/students/[username]/layout.tsx
@@ -1,5 +1,0 @@
-export default function StudentLayout({children}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return <div>{children}</div>;
-}
