+++ client/src/app/(main)/students/loading.tsx
@@ -2,9 +2,5 @@ import React from 'react';
 import { DataTableSkeleton } from '@/features/students/components/table-skeleton';
 
 export default function Loading() {
-  return (
-    <div>
-      <DataTableSkeleton columnCount={4} filterableColumnCount={2} />
-    </div>
-  );
+  return <DataTableSkeleton columnCount={4} filterableColumnCount={2} />;
 }
